Do not track entity unloads for tracker
ChunkMap#removeEntity is invoked for each entity remove, which
means that the TrackedEntity instance will always be cleared -
making our logic do nothing.

+++ src/main/java/ca/spottedleaf/moonrise/common/util/SimpleRandom.java
@@ -48,6 +48,6 @@ public final class SimpleRandom extends LegacyRandomSource {
 
         // https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
         final long value = this.advanceSeed() >>> (BITS - Integer.SIZE);
-        return (int)((value * bound) >>> Integer.SIZE);
+        return (int)((value * (long)bound) >>> Integer.SIZE);
     }
 }
